+++ frontend/components/Header.tsx
@@ -57,7 +57,7 @@ function Header() {
         <AppBar color="inherit">
           <Toolbar>
             <ResponsiveMoocLogoContainer>
-              <NextI18Next.Link href="/">
+              <NextI18Next.Link href="/" as="/">
                 <HomeLink href="/" aria-label="MOOC.fi homepage">
                   <MoocLogo alt="MOOC logo" src="../static/images/moocfi.svg" />
                   <MoocLogoText>MOOC.fi</MoocLogoText>
