+++ frontend/i18n.js
@@ -9,10 +9,13 @@ const NextI18Next = new I18Next({
   returnObjects: true,
   defaultNS: "common",
   detection: {
-    order: ["path"],
+    order: ["querystring", "path"],
+    lookupQuerystring: "lng",
     caches: [],
   },
   load: "languageOnly",
 })
+
+NextI18Next.i18n.languages = ["en", "fi", "se"]
 
 module.exports = NextI18Next
